+++ Bencodex/Codec.cs
@@ -54,6 +54,38 @@ namespace Bencodex
         public void Encode(IValue value, Stream output) =>
             Encode(value, output, offloadOptions: null);
 
+        /// <summary>Decodes an encoded value with extended format for offloaded values from
+        /// an <paramref name="input"/> stream.</summary>
+        /// <param name="input">An input stream to decode.</param>
+        /// <param name="indirectValueLoader">An optional <see cref="IndirectValue.Loader"/>
+        /// delegate invoked when offloaded values are needed.</param>
+        /// <returns>A decoded value.</returns>
+        /// <exception cref="ArgumentException">Thrown when a given
+        /// <paramref name="input"/> stream is not readable.</exception>
+        /// <exception cref="DecodingException">Thrown when a binary representation of
+        /// an <paramref name="input"/> stream is not a valid Bencodex encoding.</exception>
+        public IValue Decode(Stream input, IndirectValue.Loader? indirectValueLoader)
+        {
+            if (!input.CanRead)
+            {
+                throw new ArgumentException("The input stream cannot be read.", nameof(input));
+            }
+
+            return new Decoder(input, indirectValueLoader).Decode();
+        }
+
+        /// <summary>Decodes an encoded value with extended format for offloaded values from
+        /// a <see cref="byte"/> array.</summary>
+        /// <param name="bytes">A <see cref="byte"/> array of Bencodex encoding.</param>
+        /// <param name="indirectValueLoader">An optional <see cref="IndirectValue.Loader"/>
+        /// delegate invoked when offloaded values are needed.</param>
+        /// <returns>A decoded value.</returns>
+        /// <exception cref="DecodingException">Thrown when a <paramref name="bytes"/>
+        /// representation is not a valid Bencodex encoding.</exception>
+        [Pure]
+        public IValue Decode(byte[] bytes, IndirectValue.Loader? indirectValueLoader) =>
+            Decode(new MemoryStream(bytes, false), indirectValueLoader);
+
         /// <summary>Decodes an encoded value from an <paramref name="input"/>
         /// stream.</summary>
         /// <param name="input">An input stream to decode.</param>
@@ -63,18 +95,7 @@ namespace Bencodex
         /// <exception cref="DecodingException">Thrown when a binary
         /// representation of an <paramref name="input"/> stream is not a valid
         /// Bencodex encoding.</exception>
-        public IValue Decode(Stream input)
-        {
-            if (!input.CanRead)
-            {
-                throw new ArgumentException(
-                    "stream cannot be read",
-                    nameof(input)
-                );
-            }
-
-            return new Decoder(input).Decode();
-        }
+        public IValue Decode(Stream input) => Decode(input, indirectValueLoader: null);
 
         /// <summary>Decodes an encoded value from a
         /// <c cref="byte">Byte</c> array.</summary>
@@ -85,9 +106,6 @@ namespace Bencodex
         /// <paramref name="bytes"/> representation is not a valid Bencodex
         /// encoding.</exception>
         [Pure]
-        public IValue Decode(byte[] bytes)
-        {
-            return Decode(new MemoryStream(bytes, false));
-        }
+        public IValue Decode(byte[] bytes) => Decode(bytes, indirectValueLoader: null);
     }
 }